chore: Move name resolution/type inference from Resolver.cs into its own file (#3235)
This PR moves a large portion of `Resolver.cs` into a separate file,
`NameResolutionAndTypeInference.cs`. There are no other changes in this
PR, just the code move (and I also deleted one old method that I
happened to notice is no longer used and should have been deleted in
some previous PR).

The purpose of this move is to separate the name resolution and type
inference/checking of the Resolver from the rest of the Resolver code,
so that those pieces can be replaced (in an upcoming series of PRs).
Since the upcoming name resolution + type inference will replace the
code that this PR moves into a separate file, there's no point for the
present PR to try to improve the code that's being moved.

This code move was a rather rough cut-and-paste. There may be further
methods that should have been moved as well. If so, they will be moved
in a future PR. For now, the important thing is to at least move the
bulk of the code.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>
# Conflicts:
#	Source/DafnyCore/Resolver/Resolver.cs