refactor: Clean up resolution passes 0 and 1 (#3284)
This PR cleans up the AST traversals that occur during "pass 0" and "pass 1" of the resolver.
These now do:

* Pass 0
    - resolve names and infer types
    - check that type inference did not leave any types underspecified
- perform substitution for DefaultValueExpression, compute .ResolvedOp,
and some similar checks and AST updates as necessary to make the AST
coherent
* Pass 1
    - discover bounds
    - compute call graph
    - compute and check ghost things
    - determine tail recursion

Each "pass" really consists of several traversals of the AST. So,
"phase" or something like that would be a better name.

With this PR, the dependencies between pass 0 and pass 1 have been
sorted out, and so have the dependencies among the various things that
are happening in pass 1.

Pass 0 is the most delicate one, because it starts with an AST about
which very little is known. The general direction of the work in this PR
is to try to make later traversals be as independent of each other as
possible.

A few awkward dependencies remain between the passes. The one that comes
to mind is that the body of a prefix predicate/lemma cannot be filled
until the call-graph has been built. For this reason,
`CheckTypeInference` is run on these prefix bodies once they have been
constructed. It's possible that this could be improved by making sure
that the prefix bodies are created without a need to call
`CheckTypeInference` on them at all.

This PR also fixes some minor things:
* pretty-printing of `newtype` declarations
* scope and two-state checks for attributes
* inferred equality support for non-null iterators (issue #3273)

Fixes #3273

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>
# Conflicts:
#	Source/DafnyCore/Resolver/Resolver.cs